Implement refactor to utilize new donation configuration endpoint.

+++ app/src/main/java/org/thoughtcrime/securesms/badges/gifts/flow/GiftFlowViewModel.kt
@@ -83,7 +83,7 @@ class GiftFlowViewModel(
       onSuccess = { (giftLevel, giftBadge) ->
         store.update {
           it.copy(
-            giftLevel = giftLevel,
+            giftLevel = giftLevel.toLong(),
             giftBadge = giftBadge,
             stage = getLoadState(it, giftBadge = giftBadge)
           )
